fix: Load cluster when editing properly
By switching to using the clusterLayoutRoute, the cluster in the route context will be populated properly.
